Add new project files for Mitake.Sms.Core and update README
This commit introduces new project files for Mitake.Sms.Core, including the csproj files for the main library and example, as well as the related models and services. The detailed settings and package references are specified in these new files. The README file has also been updated, replacing "亞太電信e管家 SMS C# SDK" with "三竹